Refactor Merkle dist scripts
Having a "script" folder and a "src" folder can be confusing, especially
when code of both folders are related.

This refactoring intends to show a more clear architecture for scripts,
sorting all of them behind the src/ folder umbrella.

Also, package.json file of PRE rewards code has been deleted and all the
dependencies has been added to the main package.json (in the repo root)
in order to have a more clear vision of how the scripts work and a
better modularization of the code. This also simplifies the execution of
gem_rewards_dist script, since previously it was needed to run npm
install twice, one time for the main NPM package, and again for the
pre-rewards NPM package.

Readme file has been updated also.