Adding RandomBeaconChaosnet contract
This is a stub contract that will replace the original RandomBeacon
contract until we have off-chain clients ready. It will be used only for
the initial choasnet phases. Request of a relay entry will set the
callback contract and also exectute a callback function with an
arbitrary number instead of a "random beacon".