Detach the signing logic from the `tbtc.node` structure
Before starting the implementation of the serial signing, there is a need to
detach the signing logic from the `tbtc.node`. The `tbtc.node`'s responsibility
is listening for world's events, acting on them using the corresponding
use-case function, and influencing the world with the function's result. The
signing logic is not a fully-fledged use-case itself but is rather a task
that is performed by higher level mechanisms like sweeping, redemption and so
on. That said, detaching the signing logic from the `tbtc.node` structure
seems to be a right move.

The detachment was achieved by introducing a concept of a signing group
controller (`signingGroupController` structure). Each `tbtc.node` controls
only a part of the given wallet's signing group and the aforementioned component
encapsulates this ability. The controller is created using the `tbtc.node`'s
helper function called `createSigningGroupController`. The controller exposes
two functions. The `sign` function is responsible for signing an individual
message using the signers managed by the controller. The `signBatch` function
is supposed to sign multiple messages by calling the `sign` function according
to the defined coordination logic.