Update the protocol latch on when a single signer starts or ends the protocol
The signing group controller should cause an update of the `tbtc.node`'s
protocol latch when signers start and end the signing protocol. We achieve that
by introducing indirect callback functions that trigger the protocol latch's
functions underneath. This construction keeps the protocol latch inside the
`tbtc.node`, is safer and more flexible comparing to passing the protocol
latch directly to the signing group controller.