Extract the signature from the signer routines
Here we add the part that is responsible to extract a single signature
from all the controlled signers. We use the assumption that all controller
signers should always come up to the same result on the happy path case.
If a single signer fails or compute a different signature, something went
bad, and we should give up the whole signing.