Synchronize signing group members upon successful signing
Here we implement a component that is meant to synchronize all signing group
members once a signature is produced by the given attempt's participants.
All signing group members (either participating or omitting the given attempt)
must end up with the same signature and a common end block. This is crucial
for serial signing.