Implement Electrum connection initialization
The function initializes connection to a remote Electrum server.

TCP and SSL types of connections are supported.

Most of the servers run electrum protocol 1.4 version, hence we use it
for verification. If a server runs another version of the protocol a
warning will be issued.

A connection to an Electrum server may be dropped if no requests are
submitted, hence we implement a keep alive loop that will ping the
server to keep the connection alive.