Implement GetTransaction function
The function queries electrum for a details of the specific transaction and
converts it in a format expected by the bitcoin.Chain interface.

The simpliest solution is to use `GetTransaction` function to query electrum
as it returns the transaction details in verbose format. Unfortunatelly it's
not compatible with Esplora/Electrs implementation of ElectrumX server.
(see: https://github.com/Blockstream/electrs/pull/36)

As a workaround to cover integration with Esplora/Electrs we use `GetRawTransaction`
and deserialize the transaction.