Ignoring extracted contract addresses by Docker
The current version of Dockerfile is split to a couple of layers. The
top layers are responsible for Go code generators. However, part of this
generation include the extraction of addresses from three contracts:
TokenStaking, RandomBeacon and WalletRegistry. Addresses are written in
the files that will later be used by the client during the startup.
But the next COPY command that copies the root dir to a Docker working dir
overwrites these 3 files with empty content. To mitigate the issue, we
can exclude these 3 files from the Docker context and they won't
overwrite addresses as part of "COPY ./ $APP_DIR/" command.