Control the broadcast channel lifecycle in a separate goroutine
During the tests of the code orchestrating tECDSA signing, we realized some
buffers get full and the broadcast channel keeps writing to them even though the
receiver is no longer alive. This was fixed by introducing an additional context
in the asynchronous state machine that unregisters the handler if the state
machine (receiver) exits sooner than the context. This has fixed the problem on
the receiver side but we still need to fix the problem on the producer side.
And this is what this commit is doing.

A separate goroutine controls the lifecycle of the handler. The message handler
is removed from the channel if the context is done. This logic is placed in
a separate goroutine because the call to `handleWithRetransmission` is blocking
and we do not want to wait with removing the handler if that call blocks for
a longer period of time, especially, when the underlying buffered channel is
full.