Increase announcer's receiver buffer to 512
During the tests of the code orchestrating tECDSA signing in, we noticed the
announcer's buffer sometimes gets full and the messages are dropped with the
famous "message handler is too slow" warning. It turns out that the problem lies
in the size of the buffer used by the announcer. Before a message in which the
announcer is not interested is dropped, it is buffered in this channel. This
also includes retransmissions from the previous signing protocols because the
retransmission cache filter is scoped to the given `Recv` handler:
```
func WithRetransmissionSupport(delegate func(m net.Message)) func(m net.Message) {
  mutex := &sync.Mutex{}
  cache := make(map[string]bool)
  // (...)
}
```

The announcer's buffer size has been increased to 512 elements, just like the
buffer of the asynchronous state machine, and the problem seems to be gone based
on local machine tests.