Extract electrum requests to a requestWithRetry function
The electrum requests are pretty common and share reusable code. We
create a wrapper function for the electrum requests.

Another thing that has to be included in the request call is a context
which is specific for the request only, and should have a timeout
defined for the single request.