Reconnect if connection is down
A connection to electrum server can get down, we want to reconnect in
such case.
The `reconnectIfShutdown` function checks if the client was shutdown
(this happens when request to a server fails). If it is down the
function will reconnect to the server.
We introduced mutex to handle concurrent requests that won't overlap the
reinitialization if a request happens at the same time.

The reconnect is a part of `requestWithRetry` function which is a
wrapper for any request to the electrum server, including keep alive
pings and protocol requests.