New Merkle Distribution for Dec 1st
This is the rewards distribution held on Dec 1st 2022.

As this is a cumulative Merkle distribution, the rewards here collected
are the sum of the rewards distributed in previous Merkle distributions
plus the rewards earned in the period between the last and the current
distribution.

This includes:

- Bonus rewards for whose who were staking and had an PRE operator
  confirmed before June 1st 2022 and didn't unstaked any amount of
  tokens until July 15th 2022.
- PRE rewards for those who were staking and had an PRE operator
  confirmed between June 1st and Dec 1st 2022.
- tBTCv2 rewards for those who comply the tBTCv2 requirements between
  Sep 30th 2022 and Dec 1st 2022.

The reward allocations for this distribution is:

- PRE rewards: 75%
- tBTCv2 rewards: 25%