Simplified DKG result processing logic
We should not look at the misbehaved members from DKG result because this value
should be equal to the one returned from the DKG retry loop. The protocol
requires all members selected to the given attempt to cooperate to the very end.
This condition allows to simplify the code in the following way:

1. Result is produced. It means that everyone who was supposed to participate
   did participate to the very end.
2. Everyone agreed to the result. They did participate to the very end. We can
   shift the indexes based on information we have and register the signer. No
   snapshot needed, it's a regular registration.
3. We publish result to the chain. We sign our version with shifted indexes and
   everyone else should have the same version.

It means:
- There is no need to decideSigningMemberFate.
- We compute the final signing group with indexes shifted and members excluded
  earlier, right after the signing protocol returned the result and before we
  registered the signer.