Move BitcoinConfig out of electrum package
The `BitcoinConfig` struct is a container for configuration of different
implementations of the bitcoin integration. We cannot place it in
`electrum` package, as the package is a specific implementation.

The struct cannot be located insige the `pkg/bitcoin` package due to a
cyclic dependencies. We move it out to the main `config` package as it
seems the most clean approach. We can regiter more implementations (e.g.
Electrs).