Adjusted the memberID / memberIndex nomenclature
Adjusted the nomenclature to be about `memberIndex` not `memberID`.
See: https://github.com/keep-network/keep-core/pull/3112#discussion_r933165274

We could live with this inconsistency but we start integrating with
ECDSA smart contracts and the member ID means something different
on-chain. In the contract, member ID is the operator ID assigned by the
sortition pool.

This had to be cleaned up in the client before the integration happens.