Fix slither job by upgrading Python to `13.11.0`
The `ubuntu-latest` runner used by us to run GH Actions jobs is getting migrated
from Ubuntu `20.04` to `22.04`
(https://github.com/actions/runner-images/issues/6399). The `22.04` version does
not work with the config of the `actions/setup-python@v4` action which we uses
in the jobs running slither. We were getting following failures in workflows:
```
Version 3.8.5 was not found in the local cache
Error: Version 3.8.5 with arch x64 not found
The list of all available versions can be found here:
https://raw.githubusercontent.com/actions/python-versions/main/versions-manifest.json
```
Based on
https://github.com/actions/setup-python/issues/555#issuecomment-1337036543, we
should update Python version to `3.9.12` or higher. Or we could configure job to
run on the hardcoded Ubuntu 20.04 runner.
According to Slither documentation `Slither requires Python 3.8+ and solc, the
Solidity compiler`, so upgrading to latest Python version (`3.11.0`) should be
ok.