Run Solidity workflows also after changes to workflow files
We're changing workflows configs so that the workflows would be triggered also
on changes to their config. This way we should be able to easily see if changes
to the workflow don't break anything (note that this does not apply to jobs that
are triggered only by the `workflow_dispatch` event - as this job publishes
important packages, we need more thoughtful approach to testing the changes
there).