Downgrade `python-version` in job running Slither to `3.10.8`
In a previous commit we were trying to upgrade Python version to `3.11.0`, but
we discovered that the Slither installation does currently not work correctly
when that version is used (there's an Issue created for that, see
https://github.com/crytic/slither/issues/1466). One of the recommended solutions
is temporary downgrade to Python `3.10`. We tested that, but for some
reason, parameter was interpreted as `3.1`. Hence we're now using
`3.10.8`, which is the highest `3.10.x` version and it is available for
Ubuntu 22.04.