Split the signing done check into a listen and a wait phases
Consider the following scenario: three members in a signing group: Bob, Eve, and
Alice. They all participate in signing and are expected to confirm they are done.

Bob finishes signing at block 1000.
Eve finishes signing at block 1001.
Alice finishes signing at block 1013.

Bob starts signaling he is done at block 1000.
Eve starts signaling she is done at block 1001.
Alice starts signaling she is done at block 1013.

This all looks good except that in the previous version of the done check, Eve
was not accepting Done messages from Bob before block 1001 and Alice was not
accepting Done messages from Bob and Eve before block 1013. Eve and Alice had to
count on retransmissions.

Since we use a backoff strategy for retransmissions
(`R _ R _ _ R _ _ _ _ R _ _ _ _ _ _ _ _ R...`), they could be lucky to receive
a message early or they could not be lucky and wait for a longer period of time.

We observed this on testnet as a problem with synchronization of the
announcement phase. Even though Bob and Eve received a signing Done message from
Alice and they moved to the next message in the batch, Alice was still waiting
for a Done message from Bob and Eve and was not joining the signing of the next
message on time.

This is a graphical representation explaining the situation:
```
B: D _ D _ _ D _ _ _ _ D _ x _ _ _ _ _ _ D
E:   D _ D _ _ D _ _ _ _ D x _ _ _ _ _ _ _ D
A:                         D
```

Everyone is sending (retransmitting) their Done message at D. Bob and Eve
receive Done message from Alice at x and they continue to the next message.
Alice waits for the next 8 blocks to receive the Done message retransmission
from Eve and 7 blocks to receive the Done message retransmission from Bob.

This problem has been solved by splitting `listen` and `exchange` functions
into three pieces. The first piece is `listen` which is supposed to be called
before the signing starts. The second piece is `signalDone` which is supposed to
be called by all members participating in the given signing attempt when they
are done. Finally, all members should call `waitUntilAllDone`. This approach is
similar to the one used in the asynchronous state machine where we accept all
types of messages all the time, even though we might not be interested in some
of them at the given moment in time.

I have also added two unit tests to improve the coverage of the done check:
- cover the path when not everyone sends the done signal,
- cover the path when one signature does not match other signatures.