Temporary heartbeat mechanism
Here we introduce a temporary heartbeat mechanism that will request a heartbeat
from the active wallet every ~8 hours. A single request consists of 5 messages
that must be signed sequentially by the target wallet. A single message has
a format of a valid heartbeat and is compatible with the fraud mechanism.

Additionally, this changeset attaches to the real Bridge contract and
takes the WalletRegistry address from it, instead of taking it from the
config. Therefore, the latter option is now deprecated and should be removed
in the near future.