Improve logging of the "non-eligible for DKG result approval" path
During tests, we observed a case when DKG participants that produced the
submitted DKG result refused to take part in result approval as they
considered themselves as non-participants. So far, the problem was
not reproduced, but we are improving logs around this path to facilitate
further debugging in case that problem repeats.