Upgrade version of Node used in CI to `v18`
Previously we'we been using `v14` of Node.js. But there are several newer
versions available. The highest version supported by the `actions/setup-node@v3`
action is `v18`. We want to upgrade from `v14` across all our modules, because
this version started to cause some problems (failing `yarn upgrade` commands)
with some of the modules.

In this commit/PR we're updating workflows relating to `random-beacon`. The
update of `ecdsa` workflows will happen in a separate commit/PR.