Refactor the code responsible for dealing with DKG result hashes
Actually, we have two types of hashes and here we emphasize this fact.

The first one is used when members sign the result to show their support and
mutually verify their signatures before attempting to submit the result.
This hash is also used during the result validation to make sure group members
signed the right thing. This is the hash the client needs to compute by itself.
This type of hash is now represented by `tecdsa/dkg.ResultSignatureHash` type.

The second type of hash is the hash used internally by the contract to store
the DKG result in a gas-efficient way. The client does not need to compute it
though that hash is populated in DKG-related events that are emitted after
submission. This type is now represented by `tbtc.DKGChainResultHash`