Harden the DKG result validation logic
The DKG result validation logic used so far was vulnerable to malicious result
re-submission and chain reorgs. Here we improve it to be resistant to both
problems. First, we improve the `notifyDKGResultSubmitted` method of the
event deduplicator to consider the DKG seed and submission block while
computing the cache key. This way, the same malicious result will not be
considered as duplicate due to different block (and sometimes seed) thus
will be validated and challenged by nodes. Second, each node will confirm
the DKG state after submitting a DKG result challenge. This way, nodes will
be able to re-submit the challenge if the previous one has not changed the
DKG state (for example, due to a chain reorg that wiped out the block containing
the previous challenge transaction)