Remove `tenderly` tag for `goerli` network
We changed the version range for `@tenderly/hardhat-tenderly` in
https://github.com/keep-network/keep-core/pull/3454 (ecdsa package) and
CI build fails on tenderly verification. The tenderly verification
passes in random-beacon package because we use a fixed `1.0.12` version.
Since this is a `dapp-development` branch, we do not actually need the
tenderly verification at least for now so here we remove the `tenderly`
tag for `goerli` network to skip the tenderly verification.