Adjust deployment scripts for testnet
Here we make some adjustments to increase the flexibility of testnet
deployments. The main goal is to have a short test feedback cycle that will
allow to check multiple scenarios using different combinations of
contract's governance parameters. In order to achieve that, we must be
able to change governable parameters quickly. This change sets a 1-second
governance delay for non-mainnet networks and passes the governance over
testnet contracts to an address controlled by the dev team.