Confirm DKG start before triggering off-chain protocol
So far, the tECDSA DKG was triggered using a simple event listener acting on
`DKGStarted` events. However, small chain reorgs may cause different
`startBlock` to be received by DKG participants. That can lead to invalid
signatures over the produced DKG result as the `startBlock` is one of the
signature's components. Here we improve that by adding a confirmation mechanism
that delays the start of the off-chain protocol until the event and DKG
on-chain state is confirmed.