Test publishing of client using code from 4c6eef5a8 ref
We want to check if the client will be published correctly (with no unwanted
files in the GCR) if we use the old version of the client code (the same one
that was used on 2022 Dec 8 to publish the client without problems). We change
the name of the image (added `-wip` suffix) to avoid adding experimental files
in the bucket used for running the clients on the `keep-test` env.