New Merkle Distribution for Feb 1st
As this is a cumulative Merkle distribution, the rewards here collected
are the sum of the rewards distributed in previous Merkle distributions
plus the rewards earned in the period between the last and the current
distribution.

This includes:

- Bonus rewards for whose who were staking and had an PRE operator
confirmed before June 1st 2022 and didn't unstaked any amount of tokens
until July 15th 2022.
- PRE rewards for those who were staking and had an PRE operator
confirmed between June 1st and Feb 1st 2023.
- tBTCv2 rewards for those who comply the tBTCv2 requirements between
Sep 30th 2022 and Feb 1st 2023.

The reward allocation for this distribution is:

PRE rewards: 50%
tBTCv2 rewards: 50%