Fixing build requirement version check
It might happen that at the end of the interval when we check the
timestamp of the latest registered build version by Prometheus, it will
return a timestamp goes pass the interval. E.g. instead of Feb 1, 2023
00:00:00 it will return Feb 1, 2023 00:01:00 which results in checking
only the latest build version instead of latest and second to latest. To
mitigate this issue, we add 30min tolerance for Prometheus timestamp of
the latest registered build version.