Fixing the build version requirement
Note that this fix didn't affect the past rewards distributions because
we had only one version of keep-core client.

This change is taking into account all the instances that an operator is
running whereas the prev. version of the code took into account only the
latest operator instance, which was not correct. Here we validated
instances in a loop against allowed build versions.