Removed stubbed OnHeartbeatRequested implementation
The plan is to use WalletCoordinator contract to coordinate heartbeats.
This code will be added in the next commits. This commit removes the
stub implementation we used for the last couple of months to simulate
event triggering at fixed frequency.

I have also realized we had a typo in OnHeartbeatRequestSubmitted. Fixed
it.