Temporarily exclude `LightRelayMaintainerProxy` from binding generation process
The `LightRelayMaintainerProxy` is not yet deployed on environments other
than `development`. That means npm packages downloaded as part of
non-development deployments do not contain their deployment info and the
`make generate` process cannot extract the contract address from there which
ends with an error. We temporarily exclude the `LightRelayMaintainerProxy`
contract from the binding generation process to workaround this problem.