Add CLI command to list tBTC deposits
We added a CLI command to list the tBTC deposits. The command queries
the chain for revealed deposits and outputs their details to stdout.

To list all deposits run:
```
keep-client wallet deposits
```

Following flags can be used with the command:
- `--wallet <address>` - filter deposits only for given wallet
- `--hide-swept` - doesn't show deposits that were already swept
- `--sort-amount` - sort output by the deposit amount

The command requires `ethereum` node details to be provided in the
config.