Leverage deposits reveal blocks passed within the sweep proposal
This changeset simplifies the deposit sweep action by leveraging the
deposits reveal blocks that are now part of the deposit sweep proposal and
are set by the coordinator. This way, the action does not have to determine the
reveal block on their own. That reduces the execution time and the number
of points of failure.

Moreover, the action does no longer need to fetch the deposit request
from the `Bridge` (the `GetDepositRequest` call) because the
`PastDepositRevealedEvents` now targets just one block, and it is enough to
filter by wallet PKH. Additional filtering using the depositor address is
redundant.