Fix broadcast function in Electrum implementation
So far, the `Broadcast` method of our Electrum client tried to broadcast
the transaction as a UTF string
(`client.BroadcastTransaction(ctx, string(rawTx))`). That was wrong as the
client expects a hexadecimal string. Here we are fixing that by building
`rawTx` as `hex.EncodeToString(transaction.Serialize())` which produces
a proper hex of the raw transaction.