Optimized signing retry loop by skipping retries that cannot succeed
Think of this scenario:
1. There are three nodes: A, B, C. Each one has members in a wallet.
2. Only node A is up.
3. Request for signing is emitted on the chain (sweep or heartbeat).
4. Node A receives the event and activates the action and the signing retry loop.
5. Node A fails in the signing announcement phase. First retry, second retry,
   third retry. Nodes B and C are still offline.
6. B and C go up.
7. At this point, node A is at the fourth signing attempt.

What was happening: Nodes B and C kept executing retry attempts from 1 to 3.
They were taking less than a second because the attempt start block and attempt
and block were both in the past. We had a lot of warnings like:

2023-05-03T13:19:04.787+0200 WARN keep-tbtc tbtc/signing_loop.go:247 [member:41]
completed announcement phase for attempt [1] with minority of [1] members ready
to sign; starting next attempt (...)

Finally, they were able to catch up with node A and execute signing but we had
a lot of warnings along the way that could mislead people not being
super-familiar with how retry attempts were implemented. Also, it was not
efficient to even try to execute the signing if there was no chance the
announcement will go through.

The improvement is simple. If announcementEndBlock is in the past, we skip the
signing attempt and go to the next one. This was implemented here.

This is how this situation manifests in the logs now:

2023-05-05T23:00:20.107+0200 INFO keep-tbtc	tbtc/signing_loop.go:215 [member:40]
skipping attempt [1]; the current block is [24860] and the end block [24751] for
the announcement phase is in the past	(...)