Remove pre-signing delay from deposit sweep action
The pre-signing delay was introduced to compensate the time spent for
proposal validation. However, during testing, it turned out the time
spent for validation is negligible thus the delay can be safely removed.
Potential anomalies are supposed to be handled by the signing announcement
phase so, an additional pre-signing delay is redundant. This change should
speed up the deposit sweep action by approximately `2 min * N`, where `N` is the
number of swept deposits.