Ensure wallet state is synced between BTC and host chain before sweeping
The `Bridge` contract enforce that all BTC transactions produced by the given
wallet form a sequence and refer each other. If a wallet violated this rule,
such transactions would not be SPV-proved to the `Bridge` and would be used
to create undefeatable fraud challenges against the wallet. That leads to funds
loss and wallet punishment. The current state of things is vulnerable to this
problem as there are several unlikely corner cases that are not covered on
the client side. This changeset aims to fix that by introducing a sanity check
(`ensureWalletSyncedBetweenChains`) executed at the beginning of the sweep
action. This check ensures the wallet state on the BTC chain is correctly
reflected on the host chain `Bridge` before starting the sweep action. This way,
it makes sure the expected transaction sequence is preserved and produced
transaction will be proper from `Bridge`'s standpoint.

The `ensureWalletSyncedBetweenChains` checks whether two conditions are met:
- The wallet main UTXO registered in the host chain Bridge comes from the
  latest BTC transaction OR wallet main UTXO is unset and wallet's BTC
  transaction history is empty. This condition ensures that all expected SPV
  proofs of confirmed BTC transactions were submitted to the host chain Bridge
  thus the wallet state held known to the Bridge matches the actual state
  on the BTC chain.
- There are no pending BTC transactions in the mempool. This condition
  ensures the wallet doesn't currently perform any action on the BTC chain.
  Such a transactions indicate a possible state change in the future
  but their outcome cannot be determined at this stage so, the wallet
  should not perform new actions at the moment.