Replace flag --goerli with --testnet
We define a general `--testnet` flag for the client that replaces
the `--goerli` flag.
The flag determines the threshold network against which the client
runs and is used to resolve Ethereum and Bitcoin networks.

After we added Bitcoin support the `--goerli` flag was confusing
to use, as it relates to Ethereum only.

To run the client against the mainnet network no flag is required.
To maintain backward compatibility the `--mainnet` flag remained
available but it was marked as deprecated. We will remove the
`--mainnet` flag in the future.

Refs: https://github.com/keep-network/keep-core/issues/3579