Set Node.js version to `18.16.0`
We want to fix problem with the `yarn build` (`hardhat compile`) command
sometimes not generating expected artifacts. The problem is likely caused by the
process silently quiting, which may be related to the used version of Node (as
suggested in https://github.com/NomicFoundation/hardhat/issues/3877). We're
verifying this by first setting the version to `18.16.0`, which is the version
on which we observed the failure. If we make sure the problem is reproduceable,
we will try downgrading to see if it fixes the issue.