Read embeded electrum servers in the electrum integration tests
We want to execute the whole suite of tests for all the servers
including the embedded ones.

To do that we have to get the list from the config package, we use go:linkname
for that, instead of exposing the readEmbeddedServers function publicly.

We combine testnet and mainnet servers in the config, distinguishing it
with by bitcoin network property.

We renamed the package to `electrum_test` to overcome the cyclic
dependency problem that we have between the config and electrum
packages.

We read the servers in init function and then remove duplicates (by
URL), so we don't execute the test twice for the same server.