Log electrum server URL only when it is auto-selected
We are not logging Ethereum server URL and we expect it to be set in the
configuration.

For Electrum server, we use a similar behavior now. When it is set in
the configuration, we are not logging it. When it is auto-selected from
the servers hardcoded in the binary, we print it to logs in the config
logger.

I have also moved the log informing about Ethereum network to the config
level and added there information about Bitcoin network.