Update btc difficulty maintainer k8s config
Switch to the latest version of the keep-client released under
`v2.0.0-m3` tag.

This version has two changes affecting the k8s configuration:
- `--mainnet` flag got deprecated,
- electrum protocol is now included in the URL.

We also took a chance to define the electrum url in a dedicated config
map, and simplified the URL to `electrumx.bitcoin` which is how
kubernetes exposes the service internally.

We switched to tcp URL, as SSL is not working due to outdated
certificate.