Update dependencies to `solidity-contracts` and `random-beacon`
We have published new `@threshold-network/solidity-contract` and
`@keep-network/random-beacon` packages. Those packages no longer include
the `prepare-dependencies.sh` script from
`@threshold-network/solidity-contracts` which was causing random failures during
`yarn install`. As in some CI jobs we install the dependencies based on the
lockfile (with the `--frozen-lockfile` flag), we need to update the dependencies
in the lockfile so that the new, improved packages would be used.