Parametrize redemption transaction shape
Production code requires that the change output is always placed as the
first one in the output vector to leverage the gas optimizations made
in the Bridge implementation. However, the existing testnet transactions
we are planning to use for unit tests have the change output as the last
one. In order to mitigate that difference, we add support for both
transaction shapes.