Update CLI function for sweep proposals
Changes to support automated sweep proposals from CLI, along with the
old way of manual sweeps.

Add `deposit-sweep-max-size` flag to specfify ceiling for the deposits
count.

If args are passed to the command it will use the args as deposits to
sweep.
If the args are not provided the command will fetch deposits
automatically for the oldest unswept wallet.