Remove WalletPublicKeyHash type
As per
[comment](https://github.com/keep-network/keep-core/pull/3626#discussion_r1226587548)
@pdyraga:

We have intentionally decided to represent the public key hash as `[20]byte` in `tbtc`
package and I would rather represent it the same way here. We do not have any extra logic
related to public key hashes in this file than parsing them from string and printing them
to string. Parsing can happen in the CLI code and printing can happen inline.

The reasoning was that if we want to represent wallet pubkey hash as a separate type, then
we should do the same with refund pubkey hash, refund locktime and blinding factor. All
those types would add a lot of boilerplate code with no clear advantage.