Get deposits from two most recent wallets
Only the two most recently created wallets in the system are
sweeping. The most recent one is sweeping because it's an
active wallet. The almost-most-recently created wallet is
sweeping because it was the active wallet a short time ago and there
is a chance some deposits were revealed at the moment the active wallet
was switched or that the reveal transactions were in Ethereum mempool for a while